requirements.txt
aws-cdk.aws-apigateway==1.22.0
aws-cdk.aws-ec2==1.22.0
aws-cdk.aws-elasticsearch==1.22.0
aws-cdk.aws-iam==1.22.0
aws-cdk.aws-kinesis==1.22.0
aws-cdk.aws-lambda==1.22.0
aws-cdk.aws-lambda-event-sources==1.22.0
aws-cdk.aws-logs==1.22.0
aws-cdk.aws-s3==1.22.0
aws-cdk.core==1.22.0
